+++ release_notes.md
@@ -147,3 +147,4 @@ First release into griffon artifacts portal.
 1. Remove obsoletes methods in `SimpleJpaUtil`.
 1. Add `obfuscate` command and support `obfucasted:` string in JPA configurations.
 1. Add `isNotRowSelected` to `GlazedTable`.
+1. Remove dependency to Reflections and use `simple-jpa-annotated.txt` file to register annotated domain classes.